requirements.txt
streamlit==1.41.1
langchain==0.3.14
langchain-groq==0.2.3
langchain-core==0.3.30
langchain-community==0.3.14
huggingface-hub==0.27.1
numpy==1.26.4
pandas==2.2.3
sentence-transformers==3.3.1
transformers==4.48.0
python-dotenv==1.0.1
PyMuPDF==1.25.2
newspaper3k==0.2.8
lxml_html_clean
docarray
